wrappers/stripe-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
aio-pika
stripe
python-dotenv
grpc